feat: added index, comments and removed commented/unused code
Hooks exports have indexes to make it easier to import it in other modules. Added a bunch of comments to explain the logic behind some decisions and how the app should work once we have the implementations in place (this should be removed later). Added TODOs to track the tasks that are pending